fix test for windows
